third_party/blink/web_tests/external/wpt/html/canvas/element/manual/imagebitmap/createImageBitmap-flipY-expected.txt

This is a testharness.js-based test.
[FAIL] createImageBitmap from an HTMLVideoElement imageOrientation: "from-image", and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (5, 5) expected 247 +/- 3 but got 255
[FAIL] createImageBitmap from an HTMLVideoElement imageOrientation: "flipY", and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (5, 5) expected 28 +/- 3 but got 2
[FAIL] createImageBitmap from an HTMLVideoElement from a data URL imageOrientation: "from-image", and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (5, 5) expected 247 +/- 3 but got 255
[FAIL] createImageBitmap from an HTMLVideoElement from a data URL imageOrientation: "flipY", and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (5, 5) expected 28 +/- 3 but got 2
Harness: the test ran to completion.
